StepFourForm component created and imported in the UserForm (now with less code lines).
